Switched to renamed & refactored 'context-propagation' library.
