+++ pom.xml
@@ -52,7 +52,7 @@
 
         <!-- runtime -->
         <opentracing-api.version>0.20.2</opentracing-api.version>
-        <context-propagation.version>0.1.0</context-propagation.version>
+        <context-propagation.version>0.2.0</context-propagation.version>
 
         <!-- optional provided -->
 
@@ -231,16 +231,16 @@
         <dependency>
             <groupId>io.opentracing</groupId>
             <artifactId>opentracing-api</artifactId>
-            <version>0.20.2</version>
+            <version>${opentracing-api.version}</version>
         </dependency>
         <dependency>
             <groupId>io.opentracing</groupId>
             <artifactId>opentracing-noop</artifactId>
-            <version>0.20.2</version>
+            <version>${opentracing-api.version}</version>
         </dependency>
         <dependency>
             <groupId>nl.talsmasoftware</groupId>
-            <artifactId>concurrency</artifactId>
+            <artifactId>context-propagation</artifactId>
             <version>${context-propagation.version}</version>
         </dependency>
